requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-ollama>=0.2.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
pydantic>=2.0.0
rich>=13.0.0
python-dateutil>=2.8.0
aiohttp>=3.9.0
